+++ src/amun/common/scorer.h
@@ -44,7 +44,7 @@ class Scorer {
     virtual ~Scorer() {}
 
     virtual void Decode(const God &god, const State& in,
-                       State& out, const std::vector<size_t>& beamSizes) = 0;
+                       State& out, const std::vector<uint>& beamSizes) = 0;
 
     virtual void BeginSentenceState(State& state, size_t batchSize=1) = 0;
