+++ src/amun/fpga/best_hyps.cpp
@@ -23,7 +23,7 @@ void BestHyps::DisAllowUNK(mblas::Matrix& Prob)
   SetColumn(Prob, UNK_ID, std::numeric_limits<float>::lowest());
 }
 
-void BestHyps::FindBests(const std::vector<size_t>& beamSizes, mblas::Matrix& Probs,
+void BestHyps::FindBests(const std::vector<uint>& beamSizes, mblas::Matrix& Probs,
                std::vector<float>& outCosts,
                std::vector<unsigned>& outKeys,
                const bool isFirst)
@@ -38,7 +38,7 @@ void BestHyps::CalcBeam(
     const Words& filterIndices,
     bool returnAlignment,
     std::vector<Beam>& beams,
-    std::vector<size_t>& beamSizes
+    std::vector<uint>& beamSizes
     )
 {
   using namespace mblas;
