+++ src/amun/fpga/encoder_decoder.cpp
@@ -50,7 +50,7 @@ void EncoderDecoder::BeginSentenceState(State& state, size_t batchSize)
 }
 
 void EncoderDecoder::Decode(const God &god, const State& in,
-                   State& out, const std::vector<size_t>& beamSizes)
+                   State& out, const std::vector<uint>& beamSizes)
 {
   const EDState& edIn = in.get<EDState>();
   EDState& edOut = out.get<EDState>();
