+++ src/amun/gpu/decoder/best_hyps.h
@@ -33,7 +33,7 @@ class BestHyps : public BestHypsBase
       SetColumn(Prob, UNK_ID, std::numeric_limits<float>::lowest());
     }
 
-    void FindBests(const std::vector<size_t>& beamSizes, mblas::Matrix& Probs,
+    void FindBests(const std::vector<uint>& beamSizes, mblas::Matrix& Probs,
                    std::vector<float>& outCosts,
                    std::vector<unsigned>& outKeys,
                    const bool isFirst) {
@@ -70,7 +70,7 @@ class BestHyps : public BestHypsBase
           const Words& filterIndices,
           bool returnAlignment,
           std::vector<Beam>& beams,
-          std::vector<size_t>& beamSizes
+          std::vector<uint>& beamSizes
           )
     {
       using namespace mblas;
