+++ src/amun/gpu/decoder/encoder_decoder.cu
@@ -44,7 +44,7 @@ void EncoderDecoder::BeginSentenceState(State& state, size_t batchSize) {
   decoder_->EmptyEmbedding(edState.GetEmbeddings(), batchSize);
 }
 
-void EncoderDecoder::Decode(const God &god, const State& in, State& out, const std::vector<size_t>& beamSizes) {
+void EncoderDecoder::Decode(const God &god, const State& in, State& out, const std::vector<uint>& beamSizes) {
   const EDState& edIn = in.get<EDState>();
   EDState& edOut = out.get<EDState>();
